requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.40.0
voyageai>=0.3.0
chromadb>=0.5.0,<2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
pydantic>=2.9.0
pypdf>=5.0.0
markdown>=3.7
